Crate diemdb

This crate provides DiemDB which represents physical storage of the core Diem data structures.

It relays read/write operations on the physical storage via schemadb to the underlying Key-Value storage system, and implements diem data structures on top of it.

Modules

backup

diemsum

errors
This module defines error types used by DiemDB.

metrics

schema
This module defines representation of Diem core data structures at physical level via schemas that implement schemadb::schema::Schema.

test_helper
This module provides reusable helpers in tests.

Structs

DiemDB
This holds a handle to the underlying DB responsible for physical storage and provides APIs for access to the core Diem data structures.

Traits

GetRestoreHandler

Functions

test_save_blocks_impl