Struct diemdb::DiemDB

pub struct DiemDB { /* private fields */ }

This holds a handle to the underlying DB responsible for physical storage and provides APIs for access to the core Diem data structures.

Implementations

impl DiemDB

pub fn open<P: AsRef<Path> + Clone>(
db_root_path: P,
readonly: bool,
prune_window: Option<u64>,
rocksdb_config: RocksdbConfig
) -> Result<Self>

pub fn open_as_secondary<P: AsRef<Path> + Clone>(
db_root_path: P,
secondary_path: P,
rocksdb_config: RocksdbConfig
) -> Result<Self>

pub fn new_for_test<P: AsRef<Path> + Clone>(db_root_path: P) -> Self

This opens db in non-readonly mode, without the pruner.

pub fn update_rocksdb_properties(&self) -> Result<()>

This force the db to update rocksdb properties immediately.

pub fn get_backup_handler(&self) -> BackupHandler

Gets an instance of BackupHandler for data backup purpose.